Prototype bridge_processor and create mock data
Created a temporory tests directory that has a script that loads a
test_config for the bridge processor and then acquires mock data from
the ripple libraries on s3.

Edited the default config.yaml and the CollectionData object to point to
the bridge index on s3

Created prototype bridge_processor.py. The output still needs to be
inspected/verified. This file is called by the test script along with
create_extent_lib.

Very minor changes to the f-string formatting in a few other files to
get the code to work with my python environment. These changes can be
discarded if we get the environment quirks sorted out.