Bridge processor (#2)
* Add make_tile_index.py tools

make_tile_index.py uses the pthyon bindings to gdaltindex to make a tile
index from a directory of rasters on S3 in parallel

* Prototype bridge_processor and create mock data

Created a temporory tests directory that has a script that loads a
test_config for the bridge processor and then acquires mock data from
the ripple libraries on s3.

Edited the default config.yaml and the CollectionData object to point to
the bridge index on s3

Created prototype bridge_processor.py. The output still needs to be
inspected/verified. This file is called by the test script along with
create_extent_lib.

Very minor changes to the f-string formatting in a few other files to
get the code to work with my python environment. These changes can be
discarded if we get the environment quirks sorted out.

* Fix dry value and make bridge processing elevation unit aware

Was previously assuming dry value was 0 when it is actually the rasters
no-data value.

I also needed to convert the elevation units from meters to feet for the
bridge rasters since the terrain dem and the flood depths are in feet.

* Cleanup bridge_processor.py and test script

Remov hidden config for bridge elevation conversion and add to
config.yaml.

Redo logging and stdout to conform to the conventions used by the extent
processor.

Add clarifying comments.

General cleanup of both bridge processor and test script for conciseness
and readability.

* Streamline bridge_processor.py

More changes for code conciseness of bridge processor

* Add comment

* Draft convert bridge_processor.py to use subprocess calls

* Add multiprocessing

* Remove wrapper around bridge masking function

* Move vrt creation to reach level

VRT creation only happens once per reach now. Added explanatory
comments. Modified the BRIDGE_TILE_INDEX_PATH to use the "vsis3" prefix
since now using OGR to query.

* Explicitely convert bridge and dem CRS to depth CRS

* Add --hideNoData flag to gdal_calc

* Add bridge processor cell to process.ipynb

* Remove tests since now using notebook to test bridge processor on windows

* Revise bridge_processor

Changes in this PR:
1. Remove unecessary requirements from requirements.txt
2. Update process.ipynb to remove s3 session handling now that the
bridge rasters are accessible from the windows instance IAM
3. Update make_tile_index to make it resumable
4. Update bridge_processor so that it doesn't iterate through all
library tifs. It only works with a reaches tiffs when they actually
intersect bridges